requirements.txt
requests
google-auth
google-auth-oauthlib
google-api-python-client
google-generativeai
pyyaml
